[MIG] auth_oauth_autologin: Migration to 17.0
