+++ auth_oauth_autologin/__manifest__.py
@@ -5,7 +5,7 @@
     "name": "Auth Oauth Autologin",
     "summary": """
         Automatically redirect to the OAuth provider for login""",
-    "version": "16.0.1.0.0",
+    "version": "17.0.1.0.0",
     "license": "AGPL-3",
     "author": "ACSONE SA/NV,Odoo Community Association (OCA)",
     "maintainers": ["sbidoul"],
@@ -13,7 +13,5 @@
     "depends": ["auth_oauth"],
     "data": ["views/auth_oauth_provider.xml"],
     "demo": [],
-    "assets": {
-        "web.assets_frontend": ["auth_oauth_autologin/static/src/js/web_login.js"]
-    },
+    "assets": {"web.assets_frontend": ["auth_oauth_autologin/static/src/js/*.js"]},
 }
